requirements.txt
Flask
flask-cors
flask-bcrypt
mysql-connector-python
gunicorn
